+++ e2e/next-sandbox/pages/storage/map.tsx
@@ -1,16 +1,15 @@
-import {
-  RoomProvider,
-  useMap,
-  useRedo,
-  useUndo,
-  LiveblocksProvider,
-} from "@liveblocks/react";
+import { createRoomContext } from "@liveblocks/react";
 import { LiveMap } from "@liveblocks/client";
 import React from "react";
 import randomNumber from "../../utils/randomNumber";
 import createLiveblocksClient from "../../utils/createClient";
 
 const client = createLiveblocksClient();
+
+const { RoomProvider, useMap, useRedo, useUndo } = createRoomContext<
+  never,
+  { map: LiveMap<string, string> }
+>(client);
 
 export default function Home() {
   let roomId = "e2e-storage-map";
@@ -21,14 +20,12 @@ export default function Home() {
     }
   }
   return (
-    <LiveblocksProvider client={client}>
-      <RoomProvider
-        id={roomId}
-        initialStorage={{ map: new LiveMap<string, string>() }}
-      >
-        <Sandbox />
-      </RoomProvider>
-    </LiveblocksProvider>
+    <RoomProvider
+      id={roomId}
+      initialStorage={{ map: new LiveMap<string, string>() }}
+    >
+      <Sandbox />
+    </RoomProvider>
   );
 }