+++ examples/nextjs-live-cursors-chat/components/Cursor.tsx
@@ -12,7 +12,7 @@ export default function Cursor({ color, x, y, message }: Props) {
     <div
       className="absolute pointer-events-none top-0 left-0"
       style={{
-        transition: "transform 0.5s cubic-bezier(.17,.93,.38,1)",
+        transition: "transform 120ms linear",
         transform: `translateX(${x}px) translateY(${y}px)`,
       }}
     >