+++ examples/nextjs-live-cursors-chat/components/ReactionSelector.tsx
@@ -7,7 +7,7 @@ type Props = {
 export default function ReactionSelector({ setReaction }: Props) {
   return (
     <div
-      className="rounded-full px-2 transform -translate-x-1/2 -translate-y-1/2"
+      className="bg-white rounded-full px-2 transform -translate-x-1/2 -translate-y-1/2"
       style={{
         boxShadow:
           "0 0 0 0.5px rgba(0, 0, 0, 0.08), 0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)",