+++ examples/nextjs-threejs-shoe/next.config.js
@@ -1,9 +1,0 @@
-const withTM = require("next-transpile-modules")([
-  "three",
-  "react-three-fiber",
-  "drei",
-]);
-
-module.exports = withTM({
-  reactStrictMode: true,
-});