+++ examples/nextjs-whiteboard-advanced/src/components/SelectionBox.tsx
@@ -31,7 +31,7 @@ const SelectionBox = memo(
         <rect
           className={styles.selection}
           style={{
-            transition: isAnimated ? "all 0.1s ease" : "",
+            transition: isAnimated ? "all 120ms linear" : "",
             transform: `translate(${bounds.x}px, ${bounds.y}px)`,
           }}
           x={0}
@@ -49,7 +49,7 @@ const SelectionBox = memo(
                 cursor: "nwse-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "all 120ms linear" : "",
                 transform: `translate(${bounds.x - HANDLE_WIDTH / 2}px, ${
                   bounds.y - HANDLE_WIDTH / 2
                 }px)`,
@@ -67,7 +67,7 @@ const SelectionBox = memo(
                 cursor: "ns-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${
                   bounds.x + bounds.width / 2 - HANDLE_WIDTH / 2
                 }px, ${bounds.y - HANDLE_WIDTH / 2}px)`,
@@ -85,7 +85,7 @@ const SelectionBox = memo(
                 cursor: "nesw-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${
                   bounds.x - HANDLE_WIDTH / 2 + bounds.width
                 }px, ${bounds.y - HANDLE_WIDTH / 2}px)`,
@@ -103,7 +103,7 @@ const SelectionBox = memo(
                 cursor: "ew-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${
                   bounds.x - HANDLE_WIDTH / 2 + bounds.width
                 }px, ${bounds.y + bounds.height / 2 - HANDLE_WIDTH / 2}px)`,
@@ -121,7 +121,7 @@ const SelectionBox = memo(
                 cursor: "nwse-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${
                   bounds.x - HANDLE_WIDTH / 2 + bounds.width
                 }px, ${bounds.y - HANDLE_WIDTH / 2 + bounds.height}px)`,
@@ -139,7 +139,7 @@ const SelectionBox = memo(
                 cursor: "ns-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${
                   bounds.x + bounds.width / 2 - HANDLE_WIDTH / 2
                 }px, ${bounds.y - HANDLE_WIDTH / 2 + bounds.height}px)`,
@@ -157,7 +157,7 @@ const SelectionBox = memo(
                 cursor: "nesw-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${bounds.x - HANDLE_WIDTH / 2}px, ${
                   bounds.y - HANDLE_WIDTH / 2 + bounds.height
                 }px)`,
@@ -175,7 +175,7 @@ const SelectionBox = memo(
                 cursor: "ew-resize",
                 width: `${HANDLE_WIDTH}px`,
                 height: `${HANDLE_WIDTH}px`,
-                transition: isAnimated ? "all 0.1s ease" : "",
+                transition: isAnimated ? "transform 120ms linear" : "",
                 transform: `translate(${bounds.x - HANDLE_WIDTH / 2}px, ${
                   bounds.y - HANDLE_WIDTH / 2 + bounds.height / 2
                 }px)`,