+++ examples/redux-whiteboard/src/App.js
@@ -73,6 +73,7 @@ export default function App() {
               id={shapeId}
               shape={shape}
               selectionColor={selectionColor}
+              transition={selectedShape !== shapeId}
             />
           );
         })}
@@ -96,7 +97,7 @@ export default function App() {
   );
 }
 
-const Rectangle = ({ shape, selectionColor, id }) => {
+const Rectangle = ({ shape, selectionColor, id, transition }) => {
   const dispatch = useDispatch();
 
   return (
@@ -104,6 +105,7 @@ const Rectangle = ({ shape, selectionColor, id }) => {
       className="rectangle"
       style={{
         transform: `translate(${shape.x}px, ${shape.y}px)`,
+        transition: transition ? "transform 120ms linear" : "none",
         backgroundColor: shape.fill ? shape.fill : "#CCC",
         borderColor: selectionColor,
       }}