+++ examples/sveltekit-live-avatars/src/components/App.svelte
@@ -75,6 +75,9 @@
 
   .current_user_container {
     position: relative;
+  }
+
+  .current_user_container:not(:first-child) {
     margin-left: 2rem;
   }